⚙️ Rename LibraryItem to LibraryCard
